--- src/controller/StateChangeEvent.ts
@@ -1,9 +1,0 @@
-import { GameState } from "../GameState";
-
-export class StateChangeEvent {
-    static type = "StateChangeEventType";
-
-    static dispatch() {
-        GameState.emitter.dispatchEvent(new CustomEvent(StateChangeEvent.type));
-    }
-}
