+++ src/controller/events/MonsterAddEvent.ts
@@ -1,0 +1,7 @@
+import { createDispatch } from './eventCreation';
+
+export class MonsterAddEvent {
+    static type = "MonsterAddEventType";
+
+    static dispatch = createDispatch(MonsterAddEvent.type);
+}
