+++ src/controller/events/MonsterRemoveEvent.ts
@@ -1,0 +1,7 @@
+import { createDispatch } from './eventCreation';
+
+export class MonsterRemoveEvent {
+    static type = "MonsterRemoveEventType";
+
+    static dispatch = createDispatch(MonsterRemoveEvent.type);
+}
