+++ src/controller/events/StateChangeEvent.ts
@@ -1,0 +1,7 @@
+import { createDispatch } from './eventCreation';
+
+export class StateChangeEvent {
+    static type = "StateChangeEventType";
+
+    static dispatch = createDispatch(StateChangeEvent.type);
+}
