+++ src/controller/events/eventCreation.ts
@@ -1,0 +1,6 @@
+import { GameState } from '../../GameState'
+
+export const createDispatch = (name: string) =>
+    (detail?: object) =>
+        GameState.emitter.dispatchEvent(new CustomEvent(name, {detail}));
+
