+++ src/viewModel/EndTurnButton.ts
@@ -3,7 +3,7 @@ import * as PIXI from 'pixi.js';
 import { EndTurnAction } from '../controller/actions/EndTurnAction';
 import { IGuiElem } from './IGuiElem';
 import { GameState } from '../GameState';
-import { StateChangeEvent } from '../controller/StateChangeEvent';
+import { StateChangeEvent } from '../controller/events/StateChangeEvent';
 
 export class EndTurnButton implements IGuiElem {
   pixiElem: PIXI.Text;
